chore(phase-rag-5): apply migration 004 + RPC test suite 9/10 PASS
Final fixes to scripts/ingestion_assets/08_test_asset_queries.py after
applying migration 004:

  - All numeric/text params now carry an explicit cast. psycopg2 was
    serialising Python ints as int4 while match_assets declares
    p_min_quality as smallint, so the dispatcher rejected the call as
    UndefinedFunction. ::smallint / ::numeric / ::int casts resolve.

Migration 004 applied via scripts/apply_migrations.py. schema_migrations
also reconciled to record 003 (originally applied via the bespoke
_apply_migration_003.py script on 2026-05-24, never recorded in the
tracker) so the standard runner now has the correct view of state.

Final suite run on the pre-OGA-full dataset:
  T01 match_assets tileset semantic      PASS  got 20 (need >=3)
  T02 match_assets BGM (Phase 4 canary)  FAIL  got 0 (need >=1)  *expected
  T03 match_assets low-poly 3D character PASS  got 20 (need >=3)
  T04 match_assets HDRI semantic         PASS  got 20 (need >=1)
  T05 match_assets SFX semantic          PASS  got 20 (need >=3)
  T06 match_assets genre=visual_novel    PASS  got 10 (need >=3)
  T07 match_assets genre=retro_8bit      PASS  got 10 (need >=1)
  T08 match_loras A01 sprite             PASS  got 5  (need >=1)
  T09 match_loras A03 + multi-axis       PASS  got 5  (need >=1)
  T10 increment_asset_usage EMA          PASS  EMA correct, side
                                              effect rolled back

  PASSED: 9/10  THRESHOLD: 8/10  OVERALL: PASS

T02 is the documented audio_bgm canary — it will flip to PASS the
moment the OpenGameArt scrape (currently running in background,
~3-4h wall) lands its music records into asset_library_index.

GREEN-LIGHT for Reasoning Engine Week 1 is hereby met.

+++ scripts/ingestion_assets/08_test_asset_queries.py
@@ -73,9 +73,9 @@ def _match_assets(cur, **kw: Any) -> list[tuple]:
             p_style_pack := %(p_style_pack)s::text,
             p_genre := %(p_genre)s::text,
             p_engine := %(p_engine)s::text,
-            p_min_quality := %(p_min_quality)s,
-            p_match_threshold := %(p_match_threshold)s,
-            p_match_count := %(p_match_count)s
+            p_min_quality := %(p_min_quality)s::smallint,
+            p_match_threshold := %(p_match_threshold)s::numeric,
+            p_match_count := %(p_match_count)s::int
         )
     """
     defaults: dict[str, Any] = {
@@ -102,7 +102,7 @@ def _match_loras(cur, **kw: Any) -> list[tuple]:
             p_asset_type := %(p_asset_type)s::text,
             p_genre := %(p_genre)s::text,
             p_base_model_preferred := %(p_base_model_preferred)s::text,
-            p_match_count := %(p_match_count)s
+            p_match_count := %(p_match_count)s::int
         )
     """
     defaults: dict[str, Any] = {
